Move zstd.h to linux/zstd.h
